backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
requires-python = ">=3.14,<3.15"
dependencies = [
    "jinja2",
    "uvicorn",
    "pydantic[email]",
    "django-simpleui",
    "django-ninja-jwt",
    "psycopg[binary,pool]",
    "langchain[deepseek,community]",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
